LLM_Prompting/Deterministic_workflows/requirements.txt
langchain-ollama
pydantic
pyyaml
streamlit